@@ -7,7 +7,7 @@ import (
 
 // TestFunc is a function used in defining a new Matcher. It returns true if the value passes
 // the test or false for failure.
-type TestFunc func(value interface{}) bool
+type TestFunc func(value any) bool
 
 // DescribeTestFunc is a function used in defining a new Matcher. It returns a description of
 // the test expectation.
@@ -22,7 +22,7 @@ type DescribeTestFunc func() string
 //
 // The second parameter is the function to use for making a string description of a value of
 // the expected type.
-type DescribeFailureFunc func(value interface{}) string
+type DescribeFailureFunc func(value any) string
 
 // Matcher is a general mechanism for declaring expectations about a value. Expectations can be combined,
 // and they self-describe on failure.
@@ -43,7 +43,7 @@ func New(
 
 // Test executes the expectation for a specific value. It returns true if the value passes the
 // test or false for failure, plus a string describing the expectation that failed.
-func (m Matcher) Test(value interface{}) (pass bool, failDescription string) {
+func (m Matcher) Test(value any) (pass bool, failDescription string) {
 	if m.test(value) {
 		return true, ""
 	}
@@ -57,7 +57,7 @@ func (m Matcher) Test(value interface{}) (pass bool, failDescription string) {
 	return false, fmt.Sprintf("%s\nfull value was: %s", failureDesc, DescribeValue(value))
 }
 
-func (m Matcher) test(value interface{}) bool {
+func (m Matcher) test(value any) bool {
 	if m.testFn == nil {
 		return true
 	}
@@ -71,7 +71,7 @@ func (m Matcher) describeTest() string {
 	return m.describeTestFn()
 }
 
-func (m Matcher) describeFailure(value interface{}) string {
+func (m Matcher) describeFailure(value any) string {
 	if m.describeFailureFn != nil {
 		return m.describeFailureFn(value)
 	}
@@ -81,16 +81,16 @@ func (m Matcher) describeFailure(value interface{}) string {
 // EnsureType adds type safety to a matcher. The valueOfType parameter should be any value of the
 // expected type. The returned Matcher will guarantee that the value is of that type before calling
 // the original test function, so it is safe for the test function to cast the value.
-func (m Matcher) EnsureType(valueOfType interface{}) Matcher {
+func (m Matcher) EnsureType(valueOfType any) Matcher {
 	return New(
-		func(value interface{}) bool {
+		func(value any) bool {
 			if valueOfType != nil && (reflect.TypeOf(value) != reflect.TypeOf(valueOfType)) {
 				return false
 			}
 			return m.test(value)
 		},
 		m.describeTest,
-		func(value interface{}) string {
+		func(value any) string {
 			if valueOfType != nil && reflect.TypeOf(value) != reflect.TypeOf(valueOfType) {
 				return fmt.Sprintf("expected value of type %T, was %T", valueOfType, value)
 			}
