@@ -9,13 +9,13 @@ import (
 // case-sensitively.
 func StringContains(substring string) Matcher {
 	return New(
-		func(value interface{}) bool {
+		func(value any) bool {
 			return strings.Contains(value.(string), substring)
 		},
 		func() string {
 			return fmt.Sprintf("contains %q", substring)
 		},
-		func(interface{}) string {
+		func(any) string {
 			return fmt.Sprintf("did not contain %q", substring)
 		},
 	).EnsureType("")
@@ -24,13 +24,13 @@ func StringContains(substring string) Matcher {
 // StringHasPrefix is a matcher for string values that calls strings.HasPrefix.
 func StringHasPrefix(prefix string) Matcher {
 	return New(
-		func(value interface{}) bool {
+		func(value any) bool {
 			return strings.HasPrefix(value.(string), prefix)
 		},
 		func() string {
 			return fmt.Sprintf("starts with %q", prefix)
 		},
-		func(interface{}) string {
+		func(any) string {
 			return fmt.Sprintf("did not start with %q", prefix)
 		},
 	).EnsureType("")
@@ -39,13 +39,13 @@ func StringHasPrefix(prefix string) Matcher {
 // StringHasSuffix is a matcher for string values that calls strings.HasSuffix.
 func StringHasSuffix(suffix string) Matcher {
 	return New(
-		func(value interface{}) bool {
+		func(value any) bool {
 			return strings.HasSuffix(value.(string), suffix)
 		},
 		func() string {
 			return fmt.Sprintf("ends with %q", suffix)
 		},
-		func(interface{}) string {
+		func(any) string {
 			return fmt.Sprintf("did not end with %q", suffix)
 		},
 	).EnsureType("")
