@@ -157,7 +157,7 @@ func (s *chunkedStreamingHandlerImpl) removeChannel(channelToRemove chan []byte)
 	// This is called when the client closed the connection.
 	go func() {
 		// Consume anything else that gets sent on this channel, until it's closed, to avoid deadlock
-		for range channelToRemove {
+		for range channelToRemove { //nolint:revive // Intentionally draining the channel
 		}
 	}()
 
@@ -176,7 +176,7 @@ func (s *chunkedStreamingHandlerImpl) removeChannel(channelToRemove chan []byte)
 	close(channelToRemove)
 }
 
-func (s *chunkedStreamingHandlerImpl) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+func (s *chunkedStreamingHandlerImpl) ServeHTTP(w http.ResponseWriter, _ *http.Request) {
 	flusher, ok := w.(http.Flusher)
 	if !ok {
 		log.Println("httphelpers.ChunkedStreamingHandler can't be used with a ResponseWriter that does not support Flush")
